wip fixes

+++ include/physx.h
@@ -149,10 +149,10 @@ private:
   Chain<CREnt>* collisionsRegistry;
   long double timeScale;
 
-  int getTimeShift(int objIndex=-1);
-  void moveObjects(int& timeShift, int objIndex=-1);
-  void collideObjects(int& timeShift, int objIndex=-1);
-  void postMotion(int& timeShift, int objIndex=-1);
+  long double getTimeShift(int objIndex=-1);
+  void moveObjects(long double& timeShift, int objIndex=-1);
+  void collideObjects(long double& timeShift, int objIndex=-1);
+  void postMotion(long double& timeShift, int objIndex=-1);
 
   ObjectPhysicsMeta * getObjectCollisions(ObjectPhysicsMeta& meta, const bool& collide=true);
   int isColliding(ObjectPhysicsMeta& A, ObjectPhysicsMeta& B, const bool& collide=false);
